requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
python-dotenv
pydantic
alembic
pydantic[email]
greenlet
httpx